fix(phase-3): widen size check to Long instead of narrowing constant
Copilot PR #22 follow-up: `EntryEntity.EMBEDDING_DIMENSIONS` is `Long`
(required by `@HnswIndex.dimensions`), and the backfill dimension check
was paying a `.toInt()` conversion on every embedding to compare against
the `Int` `FloatArray.size`. Widen the size side instead — the constant
stays the single source of truth and the comparison ergonomics improve.

Generated-by: claude-opus-4-7

+++ core-storage/src/main/kotlin/dev/anchildress1/vestige/storage/VectorBackfillWorker.kt
@@ -37,7 +37,7 @@ class VectorBackfillWorker(private val boxStore: BoxStore, private val embedder:
             ensureActive()
             try {
                 val vector = embedder(entry.entryText)
-                check(vector.size == EntryEntity.EMBEDDING_DIMENSIONS.toInt()) {
+                check(vector.size.toLong() == EntryEntity.EMBEDDING_DIMENSIONS) {
                     "Embedder returned ${vector.size}-d vector; expected ${EntryEntity.EMBEDDING_DIMENSIONS}"
                 }
                 entry.vector = vector
